Redirect when logged in, optimized code
