+++ FrontEnd/src/components/GetUser.tsx
@@ -1,0 +1,33 @@
+import React, { useEffect, useState } from 'react';
+
+export default function GetUser() {
+  interface clientPrincipal {
+    userId: string;
+    userRoles: string[];
+    claims: string[];
+    identityProvider: string;
+    userDetails: string;
+  }
+
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  const [userInfo, setUserInfo] = useState<clientPrincipal>();
+
+  useEffect(() => {
+    (async () => {
+      setUserInfo(await getUserInfo());
+    })();
+  }, []);
+
+  async function getUserInfo() {
+    try {
+      const response = await fetch('/.auth/me');
+      const payload = await response.json();
+      const { clientPrincipal } = payload;
+      return clientPrincipal;
+    } catch (error) {
+      console.error('No profile could be found');
+      return undefined;
+    }
+  }
+  return userInfo;
+}
