+++ FrontEnd/src/components/Navbar.tsx
@@ -8,28 +8,19 @@ import { useNavigate } from 'react-router-dom';
 import React, { useEffect, useState } from 'react';
 import Link from '@mui/material/Link';
 import { Typography } from '@mui/material';
+import GetUser from './GetUser';
 
 export default function Navbar() {
   let navigate = useNavigate();
-  const [userInfo, setUserInfo] = useState<any>();
-
-  useEffect(() => {
-    (async () => {
-      setUserInfo(await getUserInfo());
-    })();
-  }, []);
-
-  async function getUserInfo() {
-    try {
-      const response = await fetch('/.auth/me');
-      const payload = await response.json();
-      const { clientPrincipal } = payload;
-      return clientPrincipal;
-    } catch (error) {
-      console.error('No profile could be found');
-      return undefined;
-    }
+  interface clientPrincipal {
+    userId: string;
+    userRoles: string[];
+    claims: string[];
+    identityProvider: string;
+    userDetails: string;
   }
+
+  const userInfo: clientPrincipal | undefined = GetUser();
 
   return (
     <Box sx={{ flexGrow: 1 }}>
